Reexport symbols from index
